requirements.txt
Django==1.11.7
djangorestframework==3.6.2
psycopg2==2.7.1
django-bootstrap-form==3.1
pillow==5.0.0
boto3==1.7.25
